Install grumpy-tools before building grumpy-runtime
`pip2 install grumpy-runtime` fails with the ERROR
https://github.com/grumpyhome/grumpy/issues/113#issuecomment-521527773
because it relies on `grumpy-tools`.

`grumpy-tools` packages is specified in `setup.py` as a
dependency, and it doesn't work, because for the build process to
succeed, the dependency should already exist at the time `setup.py`
is executed.

The official workaround is `pyproject.toml` from PEP-518
https://github.com/pypa/pip/issues/2381